requirement.txt
langchain
langgraph
playwright
psycopg2-binary
google-generativeai
python-dotenv
pypdf
fpdf2